Basic checkout for git backend
When the HEAD of a branch is checked out, a staged commit ID is
generated, which is used to pass it down to the BLOB backend.

In case of the Git backend, the staged ID is maintained in a hidden
".staged" file that is ignored (added to ".gitignore"). This means that
when a repo is initialized (via Init()) a default first commit is
created that handles all this.

Currently, only "HEAD" can be passed to Checkout(). This means that only
the current branch can be checked out (which as of today is 'master',
since we don't support branching yet). Also, this means that it cannot
work in detached HEAD mode, since checking out updates the working
directory and we need to handle this through git2go.

The idgenerator now implements SHA1 method. This is used by the git
backend to generate the ID for an staged commit.

Tests are also added