Simplifying by assuming we always have a POSIX staging area
By doing so, the BLOB backend can just look at the list of
symlinks/files in the working directory directly without having the
coordinator pass it to it.

This simplifies the Checkout() logic and removes CheckoutObjects()

Also, we can assume that a checkout creates a staged version by default,
i.e. no need to have a separate 'staged' command